refactor: extract `InfluxConfiguration` out of `Configuration`
