+++ src/main/kotlin/com/github/shirleh/Main.kt
@@ -7,7 +7,7 @@ import com.github.shirleh.command.CommandHandler
 import com.github.shirleh.datacollection.*
 import com.github.shirleh.datacollection.emoji.EmojiDataCollector
 import com.github.shirleh.datacollection.MessageDataCollector
-import com.github.shirleh.persistence.persistenceModule
+import com.github.shirleh.persistence.influx.influxModule
 import discord4j.core.DiscordClient
 import discord4j.core.event.domain.Event
 import discord4j.core.event.domain.VoiceStateUpdateEvent
@@ -42,7 +42,7 @@ fun main() = runBlocking<Unit> {
     }
 
     startKoin {
-        modules(mainModule, persistenceModule, administrationModule, dataCollectionModule)
+        modules(mainModule, influxModule, administrationModule, dataCollectionModule)
     }
 
     Database.connect("jdbc:sqlite:./data/data.db", "org.sqlite.JDBC")
