+++ src/main/kotlin/com/github/shirleh/persistence/influx/Module.kt
@@ -1,9 +1,10 @@
-package com.github.shirleh.persistence
+package com.github.shirleh.persistence.influx
 
 import com.influxdb.client.InfluxDBClient
 import org.koin.dsl.module
 
-val persistenceModule = module {
+val influxModule = module {
+    single { InfluxConfiguration(get()) }
     single { InfluxClientFactory.create(get()) }
     single { get<InfluxDBClient>().writeApi }
 }
